reorganize code and add feature display
